Extracted Contact class from Phone, Address and Email
