+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,29 @@
+package seedu.addressbook.data.person;
+
+/**
+ * Represents a person's contact.
+ */
+public class Contact {
+
+    protected final String value;
+    private boolean isPrivate;
+
+    public Contact(String value, boolean isPrivate) {
+        this.value = value;
+        this.isPrivate = isPrivate;
+    }
+
+    @Override
+    public String toString() {
+        return value;
+    }
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+}
